AsyncMode class removed
